requirements.txt
Django>=5.0,<6.0
gunicorn
dj-database-url
psycopg2-binary
whitenoise
pillow
djangorestframework 